feat: 게시물 추가, 삭제 시 예외 처리 적용 (#38)

+++ instragram/src/main/java/toyproject/instragram/post/controller/PostApiController.java
@@ -57,7 +57,12 @@ public class PostApiController {
     }
 
     @DeleteMapping("/{postId}")
-    public void deletePost(@PathVariable Long postId) {
+    public void deletePost(@SignIn SignInMember signInMember, @PathVariable Long postId) {
+        validateAccess(signInMember.getMemberId(), getPostOwnerId(postId));
         postService.deletePost(postId);
+    }
+
+    private Long getPostOwnerId(Long postId) {
+        return postService.getPost(postId).getMember().getId();
     }
 }
